Fix sign-compare compiler warning
The maximum boundary have to be converted to unsigned before comparing delta to it

Building CXX object tools/CMakeFiles/cbc.dir/cbc.cc.o
tools/cbc.cc:913:19: warning: comparison of integers of different signs: 'uint64_t' (aka 'unsigned long') and 'long' [-Wsign-compare]
        if (delta > std::numeric_limits<int64_t>::max()) {
            ~~~~~ ^ ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Reviewed-on: http://review.couchbase.org/77259
Tested-by: Sergey Avseyev <sergey.avseyev@gmail.com>

+++ tools/cbc.cc
@@ -910,7 +910,7 @@ ArithmeticHandler::run()
             cmd.initial = o_initial.result();
         }
         uint64_t delta = o_delta.result();
-        if (delta > std::numeric_limits<int64_t>::max()) {
+        if (delta > static_cast<uint64_t>(std::numeric_limits<int64_t>::max())) {
             throw BadArg("Delta too big");
         }
         cmd.delta = static_cast<int64_t>(delta);
